pyproject.toml
[project]
name = "hardboiled"
version = "0.1.0"
description = "Minimal Python static site generator library"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Hugo", email = "hugo@example.com" }]
keywords = ["static-site-generator", "jinja2", "ssg"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "jinja2>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "black>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hardboiled"]

[tool.black]
line-length = 88
unstable = true

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "C90", "D205", "D212"]
ignore = ["F402", "E501"]

[tool.ruff.format]
indent-style = "space"

[tool.ruff.lint.mccabe]
max-complexity = 13

[tool.ruff.lint.isort]
force-single-line = true

[tool.pytest.ini_options]
testpaths = ["tests"]
